Modified BIdListItem
